+++ src/components/codeEditer/CodeEditer.js
@@ -1,25 +1,19 @@
-import { useState, useEffect } from 'react';
-import styled from 'styled-components';
+import { useEffect, useState } from 'react';
+import { decode, encode } from 'js-base64';
 import Editor from '@monaco-editor/react';
-import { encode, decode } from 'js-base64';
-import { postCode } from '../../api/postCode';
-import runIcon from '../../assets/run.svg';
+import styled from 'styled-components';
+import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels';
 
-function CodeEditer({ socket }) {
+import ResultContainer from './ResultContainer';
+import { postCode } from '../../api/postCode';
+
+function CodeEditer({ code, handleEditorChange, readOnly }) {
   const [runResponse, setRunResponse] = useState({});
   const [result, setResult] = useState('');
-  const [code, setCode] = useState('// 코드를 입력해주세요');
   const [isError, setIsError] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
   // eslint-disable-next-line camelcase
   const { memory, status, stderr, stdout, time } = runResponse || {};
-
-  const handleEditorChange = (e) => {
-    if (socket) {
-      socket.emit('code', e);
-      setCode(e);
-    }
-  };
 
   const runCode = async () => {
     setIsLoading(true);
@@ -41,91 +35,44 @@ function CodeEditer({ socket }) {
     }
   }, [runResponse]);
 
-  useEffect(() => {
-    if (socket) {
-      socket.on('code', (data) => {
-        setCode(data);
-      });
-    }
-  }, [socket]);
-
   return (
-    <>
-      <Editor
-        value={code}
-        height="50vh"
-        language="javascript"
-        theme="vs-dark"
-        options={{
-          inlineSuggest: true,
-          fontSize: '16px',
-          formatOnType: true,
-          autoClosingBrackets: true,
-          minimap: { scale: 10 },
-        }}
-        onChange={(e) => handleEditorChange(e)}
-      />
-      <ResultContainer>
-        <SubmissionsContainer>
-          <RunButton type="button" onClick={() => runCode()} disabled={isLoading}>
-            {isLoading ? (
-              <l-ring-2 size="25" stroke="5" stroke-length="0.25" bg-opacity="0.1" speed="0.8" color="white" />
-            ) : (
-              <img src={runIcon} alt="실행 아이콘" />
-            )}
-            Run
-          </RunButton>
-          <Submissions>
-            <Submission>memory: {memory}kb</Submission>
-            <Submission>run time: {time}s</Submission>
-            <Submission>status: {status?.description}</Submission>
-          </Submissions>
-        </SubmissionsContainer>
-        <Result $isError={isError}>{result || ''}</Result>
-      </ResultContainer>
-    </>
+    <PanelGroup autoSaveId="example" direction="vertical" style={{ height: '100vh' }}>
+      <Panel>
+        <Editor
+          value={code}
+          height="100%"
+          language="javascript"
+          theme="vs-dark"
+          options={{
+            inlineSuggest: true,
+            fontSize: '16px',
+            formatOnType: true,
+            autoClosingBrackets: true,
+            minimap: { scale: 10 },
+            readOnly,
+          }}
+          onChange={handleEditorChange}
+        />
+      </Panel>
+      <StyledHandle />
+      <Panel>
+        <ResultContainer
+          runCode={runCode}
+          isLoading={isLoading}
+          isError={isError}
+          result={result}
+          memory={memory}
+          time={time}
+          status={status}
+        />
+      </Panel>
+    </PanelGroup>
   );
 }
 
 export default CodeEditer;
 
-const ResultContainer = styled.div`
-  width: 100%;
-  height: calc(100% - 50vh);
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  position: fixed;
-  bottom: 0;
-  font-family: Consolas, 'Courier New', monospace;
-`;
-const RunButton = styled.button`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 14px 20px;
-  border-radius: 8px;
-  color: white;
-  ${({ theme }) => theme.typographies.BUTTON_TXT}
-  background-color: ${({ theme, disabled }) => (disabled ? theme.colors.GRAY : theme.colors.GREEN_2)};
-`;
-const SubmissionsContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  color: ${({ theme }) => theme.colors.WHITE};
-`;
-const Submissions = styled.div``;
-const Submission = styled.p``;
-const Result = styled.div`
-  height: calc(100% - 62px);
-  white-space: pre-wrap;
-  word-break: break-word;
-  overflow-wrap: break-word;
-  overflow-y: scroll;
-  ${({ theme }) => theme.typographies.DEFAULT_TXT};
-  color: ${({ theme, $isError }) => ($isError ? theme.colors.RED_2 : theme.colors.WHITE)};
-  line-height: 1.5;
-  letter-spacing: normal;
-  background-color: #1c1b1a;
+const StyledHandle = styled(PanelResizeHandle)`
+  background-color: ${({ theme }) => theme.colors.DARK_GRAY};
+  height: 10px;
 `;
